feat(ir::exec): virtual machine for execution
